refactor: centralize TypeScript interfaces in types.ts
- Move all component interfaces to centralized types.ts file
- Remove duplicate interface definitions from component files
- Add JSDoc comments for better documentation
